ASYLUM: fix more GCC warnings (overflow in implicit constant conversion)
- minor whitespace cleanup as well

git-svn-id: http://asylumengine.googlecode.com/svn/trunk@482 0bfb4aae-4ea4-11de-8d8d-752d95cf3e3c

+++ engines/asylum/resources/actionlist.cpp
@@ -935,7 +935,7 @@ IMPLEMENT_OPCODE(ChangeMusicById) {
 //////////////////////////////////////////////////////////////////////////
 // Opcode 0x33
 IMPLEMENT_OPCODE(StopMusic) {
-	_scene->vm()->sound()->changeMusic(_scene->getResourcePack(), kResourceMusic_FFFFFD66, 0);
+	_scene->vm()->sound()->changeMusic(_scene->getResourcePack(), (int)kResourceMusic_FFFFFD66, 0);
 }
 
 //////////////////////////////////////////////////////////////////////////
